+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Set favicon from theme provider according to overrides
+
+### Changed
+
+- Theme provider optimizations
+
 ## 0.6.0 - 2024-05-28
 ### Added
 - `v0.0.1` GridItem layout